requirements.txt
pandas>=2.2.3
numpy>=2.1.0
duckdb>=0.10.0
streamlit>=1.32.0
google-genai>=0.3.0
pydantic>=2.6.4
python-dotenv>=1.0.1
PyPDF2>=3.0.0
